Add files via upload
Identify the influence and importance of various factors on charges of health insurance. Performing Regression analysis to study the effects of the health of the patients on the insurance charges.With the help of REGRESSION ANALYSIS, we have envisioned the optimization and efficiency boosting for the health insurance system. To incorporate machine learning and health insurance sector we will create Regression Models.